database-migration/requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.0
colorama==0.4.6
tabulate==0.9.0
